Create new components from VulnerabilityScan.tsx
