+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -1,40 +1,30 @@
-import React, { useState, useEffect, useMemo, useRef } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
 import { withSearch } from '@elastic/react-search-ui';
-import {
-  Box,
-  BoxProps,
-  Button,
-  Card,
-  CardContent,
-  Grid,
-  CircularProgress,
-  Link,
-  Stack,
-  Typography
-} from '@mui/material';
-import { alpha, useTheme } from '@mui/material/styles';
-import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
-import KeyMetrics from './KeyMetrics';
-import ScanningSummary from './ScanningSummary';
-import InfoLabel from 'components/Dashboard/InfoLabel';
-import { LoginBlockedDialog } from 'components/LoginBlockedDialog';
-import { UpdateStateForm } from 'components/Register';
-import { VulnScanDataTransformed } from 'types/vuln-scan-stats';
+import { Box, BoxProps, Grid, CircularProgress, Link } from '@mui/material';
+// Context & hooks
 import { ContextType, useAuthContext } from 'context';
+import useFirstLoginPopup from 'hooks/useFirstLoginPopup';
+import { useVulnScanData } from 'hooks/useVulnScanData';
+// Constants & types
 import { ORGANIZATION_FILTER_KEY } from 'components/FilterDrawer/VSDashRegionAndOrgFilters';
 import { OrganizationShallow } from 'components/FilterDrawer/RegionAndOrganizationFilters';
-import isDataEmpty, {
-  transformVulnScanData,
-  NO_DATA_FALLBACK_LABEL
-} from 'utils/transformVulnScanData';
+// Shared components
+import { LoginBlockedDialog } from 'components/LoginBlockedDialog';
+import { UpdateStateForm } from 'components/Register';
+import InfoLabel from 'components/Dashboard/InfoLabel';
+import NoDataMessage from 'components/Dashboard/NoDataMessage';
+import FirstLoginPopup from 'components/Dialog/FirstLoginPopup';
+import InvitePendingCard from 'components/Dialog/InvitePendingCard';
+// Local dashboard components
+import KeyMetrics from './KeyMetrics';
+import ScanningSummary from './ScanningSummary';
 import RiskyOpenServices from './RiskyOpenServices';
 import TopVulnerableHosts from './TopVulnerableHosts';
 import TopVulnerabilities from './TopVulnerabilities';
 import SeverityByProminence from './SeverityByProminence';
+// Utils & assets
+import isDataEmpty from 'utils/transformVulnScanData';
 import infoIconContent from './infoIconContent.json';
-import NoDataMessage from 'components/Dashboard/NoDataMessage';
-import useFirstLoginPopup from 'hooks/useFirstLoginPopup';
-import FirstLoginPopup from 'components/Dialog/FirstLoginPopup';
 
 export interface VulnSeverities {
   label: string;
@@ -47,38 +37,70 @@ export interface WidgetProps extends BoxProps {
   children?: React.ReactNode;
 }
 
-const InitialVSData: VulnScanDataTransformed = {
-  vulnScanSummary: [],
-  vulnScanKeyMetrics: [],
-  detectedServicesKeyMetrics: [],
-  detectedHostsKeyMetrics: [],
-  detectedHostsTop5VulnerableHosts: [],
-  topVulnerabilities: [],
-  topKevVulnerabilities: [],
-  riskyServices: [],
-  severityByProminence: []
+const tooltipContentJson = infoIconContent.infoIconContent;
+
+const panelStyles = {
+  panel: {
+    border: '1px solid',
+    borderRadius: '4px',
+    borderColor: 'neutrals.main',
+    px: 3,
+    pb: 3,
+    backgroundColor: 'neutrals.white'
+  }
+} as const;
+
+const keyMetricsStyle = {
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center'
 };
 
-const tooltipContentJson = infoIconContent.infoIconContent;
+const PageSection = ({ children }: { children: React.ReactNode }) => {
+  const { user } = useAuthContext();
+  const { show: showFirstLoginPopup, close: handleCloseFirstLoginPopup } =
+    useFirstLoginPopup(user ?? null);
+  return (
+    <Box
+      sx={{
+        maxWidth: '1152px',
+        margin: 'auto',
+        px: { xs: 0, sm: 0.5, md: 1, lg: 1, xl: 0 },
+        pb: 6,
+        minHeight: '100vh'
+      }}
+    >
+      <Grid container mb={0}>
+        <Grid size={{ xs: 12 }} sx={{ mt: 4 }}>
+          <InfoLabel
+            label="Vulnerability Scanning Dashboard"
+            typographyVariant="h1"
+            viewDetails
+            tooltipContentJson={tooltipContentJson}
+          />
+        </Grid>
+      </Grid>
+      <FirstLoginPopup
+        showFirstLoginPopup={showFirstLoginPopup}
+        handleCloseFirstLoginPopup={handleCloseFirstLoginPopup}
+      />
+      {children}
+    </Box>
+  );
+};
 
 export const VulnerabilityScan: React.FC<ContextType> = (props) => {
   const { filters, removeFilter } = props;
   const {
-    apiPost,
-    apiPut,
     currentOrganization,
     user,
     apiGet,
     userMustSign,
     logout,
-    isLoggingOut,
-    setUser
+    isLoggingOut
   } = useAuthContext();
 
   const [isUpdateStateFormOpen, setIsUpdateStateFormOpen] = useState(false);
-  const theme = useTheme();
-  const [vulnScanData, setVulnScanData] =
-    useState<VulnScanDataTransformed>(InitialVSData);
 
   const [isLoginBlockedDialogOpen, setIsLoginBlockedDialogOpen] =
     useState(false);
@@ -95,9 +117,6 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
       }
     }
   }, [user, isLoggingOut]);
-
-  const { show: showFirstLoginPopup, close: handleCloseFirstLoginPopup } =
-    useFirstLoginPopup(user ?? null, apiPut, apiGet, setUser);
 
   useEffect(() => {
     const fetchAndCheckMaintenance = async () => {
@@ -144,9 +163,6 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
     };
   }, []);
 
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
   useEffect(() => {
     filters.forEach((filter) => {
       removeFilter(filter.field, filter.values[0], filter.type);
@@ -174,99 +190,19 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
     ? organizationFilter?.values[0]?.name
     : currentOrganization?.name;
 
-  useEffect(() => {
-    const selectedOrgId = orgId ? orgId : currentOrganization?.id;
-    if (!selectedOrgId) {
-      setError(
-        "Please join an organization to be shown your organization's vulnerability scan data."
-      );
-      setVulnScanData(InitialVSData);
-      return;
-    }
-    const fetchVSScan = async () => {
-      setLoading(true);
-      setError(null);
-      try {
-        const response = await apiPost('/stats/trends', {
-          body: {
-            filters: {
-              organization_id: selectedOrgId,
-              sources: ['vs', 'port', 'port_service', 'host'],
-              enhanced_data: false
-            }
-          }
-        });
-
-        const isEmpty =
-          !response?.host_summaries?.length &&
-          !response?.port_scan_summaries?.length &&
-          !response?.port_scan_service_summaries?.length &&
-          !response?.vuln_scan_summaries?.length;
-
-        if (!response || isEmpty) {
-          setVulnScanData(InitialVSData);
-          setError(
-            'No recent scan data was found for the selected organization.'
-          );
-          return;
-        }
-
-        const transformed = transformVulnScanData(response);
-
-        // If transform hit the 4th fallback, show the big NoDataMessage panel
-        const vsLabel = transformed.vulnScanSummary[0]?.vulnerabilityScan;
-        if (vsLabel === NO_DATA_FALLBACK_LABEL) {
-          setVulnScanData(InitialVSData);
-          setError('NO_DATA');
-          return;
-        }
-
-        setVulnScanData(transformed);
-      } catch (err: any) {
-        console.error(err);
-        setError(
-          err.message +
-            '. Failed to load vulnerability scan data. See the console log for more details.'
-        );
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchVSScan();
-  }, [apiPost, currentOrganization?.id, orgId]);
-
-  const keyMetricsStyle = {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center'
-  };
-
-  const header = (
-    <Grid container mb={0}>
-      <Grid size={{ xs: 12 }} sx={{ mt: 4 }}>
-        <InfoLabel
-          label="Vulnerability Scanning Dashboard"
-          typographyVariant="h1"
-          viewDetails
-          tooltipContentJson={tooltipContentJson}
-        />
-      </Grid>
-    </Grid>
-  );
-
-  // --- Page body selection (no early returns) ---
-  let content: React.ReactNode = null;
+  const {
+    data: vulnScanData,
+    loading,
+    error
+  } = useVulnScanData(orgId ? orgId : currentOrganization?.id);
 
   if (isUpdateStateFormOpen) {
-    content = (
+    return (
       <UpdateStateForm
         open={isUpdateStateFormOpen}
         user_id={user?.id ?? ''}
         onClose={async () => {
           setIsUpdateStateFormOpen(false);
-
-          // Re-fetch updated user to prevent false popup
           const updatedUser = await apiGet('/users/me');
           if (updatedUser?.state && user?.user_type !== 'globalAdmin') {
             const notifications = await apiGet('/notifications');
@@ -285,8 +221,9 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
         }}
       />
     );
-  } else if (isLoginBlockedDialogOpen && maintenanceNotification) {
-    content = (
+  }
+  if (isLoginBlockedDialogOpen && maintenanceNotification) {
+    return (
       <LoginBlockedDialog
         open={isLoginBlockedDialogOpen}
         message={maintenanceNotification.message}
@@ -296,96 +233,31 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
         }}
       />
     );
-  } else if (user?.invite_pending) {
-    content = (
-      <Box
-        sx={{
-          position: 'fixed',
-          top: 0,
-          left: 0,
-          width: '100vw',
-          height: '100vh',
-          backgroundColor: alpha(theme.palette.neutrals.main, 0.7),
-          zIndex: 1300,
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center'
-        }}
-      >
-        <Card style={{ maxWidth: 400, textAlign: 'center' }}>
-          <CardContent>
-            <CheckCircleOutlineIcon
-              sx={{
-                fontSize: 48,
-                color: (theme) => theme.palette.primary.dark,
-                mb: 1
-              }}
-            />
-            <Typography variant="body1" sx={{ mb: 2 }}>
-              Request sent! We will notify you by email once your account is
-              approved.
-            </Typography>
-            <Button
-              variant="contained"
-              sx={{
-                backgroundColor: (theme) => theme.palette.primary.dark,
-                color: '#fff',
-                '&:hover': {
-                  backgroundColor: (theme) => theme.palette.primary.main
-                }
-              }}
-              onClick={logout}
-            >
-              Logout
-            </Button>
-          </CardContent>
-        </Card>
-      </Box>
-    );
-  } else if (loading) {
-    content = (
-      <Stack
-        sx={{ maxWidth: '1152px', margin: 'auto', height: '100vh' }}
-        spacing={6}
-      >
-        {header}
+  }
+  if (user?.invite_pending) {
+    return <InvitePendingCard />;
+  }
+  if (loading) {
+    return (
+      <PageSection>
         <CircularProgress />
-      </Stack>
+      </PageSection>
     );
-  } else if (error || isDataEmpty(vulnScanData)) {
-    // ⬇️ Force the "standard" copy ONLY for the 4th-fallback case
+  }
+  if (error || isDataEmpty(vulnScanData)) {
     const noDataUserType =
       error === 'NO_DATA' ? 'standard' : user?.user_type || 'standard';
-
-    content = (
-      <Box
-        sx={{
-          maxWidth: '1152px',
-          margin: 'auto',
-          px: { xs: 0, sm: 0.5, md: 1, lg: 1, xl: 0 }
-        }}
-        pb={6}
-        height={'100vh'}
-      >
-        {header}
+    return (
+      <PageSection>
         <NoDataMessage userType={noDataUserType} />
-      </Box>
+      </PageSection>
     );
   } else if (
     vulnScanData.vulnScanSummary[0]?.assetsOwned === 0 ||
     vulnScanData.vulnScanSummary[0]?.hostsScanned === 0
   ) {
-    content = (
-      <Box
-        sx={{
-          maxWidth: '1152px',
-          margin: 'auto',
-          px: { xs: 0, sm: 0.5, md: 1, lg: 1, xl: 0 }
-        }}
-        pb={6}
-        height={'100vh'}
-      >
-        {header}
+    return (
+      <PageSection>
         <NoDataMessage
           userType={user?.user_type || 'standard'}
           headerMsg={`There is no data available for ${
@@ -405,23 +277,14 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
             )
           }
         />
-      </Box>
+      </PageSection>
     );
   } else if (
     vulnScanData.vulnScanSummary[0]?.recentlyEnrolled &&
     vulnScanData.vulnScanSummary[0]?.hostsScanned === 0
   ) {
-    content = (
-      <Box
-        sx={{
-          maxWidth: '1152px',
-          margin: 'auto',
-          px: { xs: 0, sm: 0.5, md: 1, lg: 1, xl: 0 }
-        }}
-        pb={6}
-        height={'100vh'}
-      >
-        {header}
+    return (
+      <PageSection>
         <NoDataMessage
           userType={user?.user_type || 'standard'}
           headerMsg={
@@ -437,200 +300,142 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
             )
           }
         />
-      </Box>
+      </PageSection>
     );
-  } else {
-    // main dashboard
-    content = (
+  }
+  return (
+    <PageSection>
       <Box
-        sx={{
-          maxWidth: '1152px',
-          margin: 'auto',
-          px: { xs: 0, sm: 0.5, md: 1, lg: 1, xl: 0 },
-          pb: 3
-        }}
-        minHeight="100vh"
+        justifyContent="flex-start"
+        sx={{ ...panelStyles.panel, mt: 4, mb: 3 }}
       >
-        {header}
-        <Box
-          border="1px solid"
-          borderRadius="4px"
-          borderColor="neutrals.main"
-          justifyContent="flex-start"
-          sx={{ backgroundColor: 'neutrals.white', px: 3, pb: 3, mt: 4, mb: 3 }}
+        <Grid
+          container
+          direction="row"
+          alignItems="center"
+          justifyContent="center"
         >
-          <Grid
-            container
-            direction="row"
-            alignItems="center"
-            justifyContent="center"
-          >
-            <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
-              <InfoLabel
-                label="Latest Scanning Summary"
-                tooltipContentJson={tooltipContentJson}
-              />
-            </Grid>
-            <Grid size={{ md: 12, lg: 5.5 }}>
-              <ScanningSummary data={vulnScanData.vulnScanSummary} />
-            </Grid>
-            <Grid size={{ md: 12, lg: 6.5 }} sx={keyMetricsStyle}>
-              <KeyMetrics
-                data={vulnScanData.vulnScanKeyMetrics}
-                linkData={{ orgId, orgName }}
-              />
-            </Grid>
+          <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
+            <InfoLabel
+              label="Latest Scanning Summary"
+              tooltipContentJson={tooltipContentJson}
+            />
           </Grid>
-        </Box>
-        <Box
-          border="1px solid"
-          borderRadius="4px"
-          borderColor="neutrals.main"
-          sx={{ backgroundColor: 'neutrals.white', px: 3, pb: 3, mb: 3 }}
+          <Grid size={{ md: 12, lg: 5.5 }}>
+            <ScanningSummary data={vulnScanData.vulnScanSummary} />
+          </Grid>
+          <Grid size={{ md: 12, lg: 6.5 }} sx={keyMetricsStyle}>
+            <KeyMetrics
+              data={vulnScanData.vulnScanKeyMetrics}
+              linkData={{ orgId, orgName }}
+            />
+          </Grid>
+        </Grid>
+      </Box>
+      <Box sx={{ ...panelStyles.panel, mb: 3 }}>
+        <Grid
+          container
+          direction="row"
+          alignItems="center"
+          justifyContent="space-between"
         >
-          <Grid
-            container
-            direction="row"
-            alignItems="center"
-            justifyContent="space-between"
-          >
-            <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
-              <InfoLabel
-                label="Known Exploited and Other Detected Vulnerabilities"
-                viewDetails
-                link="/inventory/vulnerabilities"
-                stateVariables={{ orgName, orgId, scanType: 'nessus' }}
-                tooltipContentJson={tooltipContentJson}
-              />
-            </Grid>
+          <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
+            <InfoLabel
+              label="Known Exploited and Other Detected Vulnerabilities"
+              viewDetails
+              link="/inventory/vulnerabilities"
+              stateVariables={{ orgName, orgId, scanType: 'nessus' }}
+              tooltipContentJson={tooltipContentJson}
+            />
           </Grid>
-          <Grid container spacing={3}>
-            <Grid size={{ xs: 12, lg: 6 }}>
-              <SeverityByProminence
-                data={vulnScanData.severityByProminence}
-                orgName={orgName}
-                orgId={orgId}
-              />
-            </Grid>
-            <Grid size={{ xs: 12, lg: 6 }}>
-              <TopVulnerabilities
-                allData={vulnScanData.topVulnerabilities}
-                kevData={vulnScanData.topKevVulnerabilities}
-                orgName={orgName}
-                orgId={orgId}
-              />
-            </Grid>
-          </Grid>
-        </Box>
-        <Grid container direction="row" justifyContent="center" spacing={3}>
+        </Grid>
+        <Grid container spacing={3}>
           <Grid size={{ xs: 12, lg: 6 }}>
-            <Grid
-              container
-              direction="column"
-              paddingBottom={2}
-              border="1px solid"
-              borderRadius="4px"
-              borderColor="neutrals.main"
-              sx={{ backgroundColor: 'neutrals.white', px: 3, pb: 3 }}
-            >
-              <Grid container direction="row" alignItems="center">
-                <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
-                  <InfoLabel
-                    label="Detected Services"
-                    tooltipContentJson={tooltipContentJson}
-                  />
-                </Grid>
-              </Grid>
-              <Grid container direction="column" spacing={2}>
-                <Grid
-                  size={{ md: 12 }}
-                  sx={{
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center'
-                  }}
-                >
-                  <KeyMetrics data={vulnScanData.detectedServicesKeyMetrics} />
-                </Grid>
-                <Grid size={{ xs: 12 }}>
-                  <Box
-                    border="1px solid"
-                    borderRadius="4px"
-                    borderColor="neutrals.light"
-                    sx={{ px: 3, pb: 2, pt: 1 }}
-                    minHeight="435px"
-                  >
-                    <InfoLabel
-                      label="Potentially Risky Services"
-                      typographyVariant="h3"
-                      headingLevel="h3"
-                      tooltipContentJson={tooltipContentJson}
-                    />
-                    <RiskyOpenServices
-                      data={vulnScanData.riskyServices}
-                      detectedServices={
-                        vulnScanData.detectedServicesKeyMetrics.find(
-                          (m) => m.title === 'Detected Services'
-                        )?.value ?? 0
-                      }
-                    />
-                  </Box>
-                </Grid>
-              </Grid>
-            </Grid>
+            <SeverityByProminence
+              data={vulnScanData.severityByProminence}
+              orgName={orgName}
+              orgId={orgId}
+            />
           </Grid>
           <Grid size={{ xs: 12, lg: 6 }}>
-            <Grid
-              container
-              direction="column"
-              border="1px solid"
-              borderRadius="4px"
-              borderColor="neutrals.main"
-              sx={{ backgroundColor: 'neutrals.white', px: 3, pb: 3 }}
-            >
-              <Grid container direction="row" alignItems="center">
-                <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
+            <TopVulnerabilities
+              allData={vulnScanData.topVulnerabilities}
+              kevData={vulnScanData.topKevVulnerabilities}
+              orgName={orgName}
+              orgId={orgId}
+            />
+          </Grid>
+        </Grid>
+      </Box>
+      <Grid container direction="row" justifyContent="center" spacing={3}>
+        <Grid size={{ xs: 12, lg: 6 }}>
+          <Grid container direction="column" sx={{ ...panelStyles.panel }}>
+            <Grid container direction="row" alignItems="center">
+              <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
+                <InfoLabel
+                  label="Detected Services"
+                  tooltipContentJson={tooltipContentJson}
+                />
+              </Grid>
+            </Grid>
+            <Grid container direction="column" spacing={2}>
+              <Grid size={{ md: 12 }} sx={keyMetricsStyle}>
+                <KeyMetrics data={vulnScanData.detectedServicesKeyMetrics} />
+              </Grid>
+              <Grid size={{ xs: 12 }}>
+                <Box
+                  border="1px solid"
+                  borderRadius="4px"
+                  borderColor="neutrals.light"
+                  sx={{ px: 3, pb: 2, pt: 1 }}
+                  minHeight="435px"
+                >
                   <InfoLabel
-                    label="Detected Hosts"
-                    viewDetails
-                    link="/inventory/domains"
-                    stateVariables={{ orgName, orgId }}
+                    label="Potentially Risky Services"
+                    typographyVariant="h3"
+                    headingLevel="h3"
                     tooltipContentJson={tooltipContentJson}
                   />
-                </Grid>
-              </Grid>
-              <Grid container direction="column" spacing={2}>
-                <Grid
-                  size={{ md: 12 }}
-                  sx={{
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center'
-                  }}
-                >
-                  <KeyMetrics data={vulnScanData.detectedHostsKeyMetrics} />
-                </Grid>
-                <Grid size={{ xs: 12 }}>
-                  <TopVulnerableHosts
-                    data={vulnScanData.detectedHostsTop5VulnerableHosts}
+                  <RiskyOpenServices
+                    data={vulnScanData.riskyServices}
+                    detectedServices={
+                      vulnScanData.detectedServicesKeyMetrics.find(
+                        (m) => m.title === 'Detected Services'
+                      )?.value ?? 0
+                    }
                   />
-                </Grid>
+                </Box>
               </Grid>
             </Grid>
           </Grid>
         </Grid>
-      </Box>
-    );
-  }
-
-  return (
-    <>
-      <FirstLoginPopup
-        showFirstLoginPopup={showFirstLoginPopup}
-        handleCloseFirstLoginPopup={handleCloseFirstLoginPopup}
-      />
-      {content}
-    </>
+        <Grid size={{ xs: 12, lg: 6 }}>
+          <Grid container direction="column" sx={{ ...panelStyles.panel }}>
+            <Grid container direction="row" alignItems="center">
+              <Grid size={{ xs: 12 }} sx={{ my: 1.5 }}>
+                <InfoLabel
+                  label="Detected Hosts"
+                  viewDetails
+                  link="/inventory/domains"
+                  stateVariables={{ orgName, orgId }}
+                  tooltipContentJson={tooltipContentJson}
+                />
+              </Grid>
+            </Grid>
+            <Grid container direction="column" spacing={2}>
+              <Grid size={{ md: 12 }} sx={keyMetricsStyle}>
+                <KeyMetrics data={vulnScanData.detectedHostsKeyMetrics} />
+              </Grid>
+              <Grid size={{ xs: 12 }}>
+                <TopVulnerableHosts
+                  data={vulnScanData.detectedHostsTop5VulnerableHosts}
+                />
+              </Grid>
+            </Grid>
+          </Grid>
+        </Grid>
+      </Grid>
+    </PageSection>
   );
 };
